Prep for GraphQL

+++ payment_allocation/database.py
@@ -14,25 +14,27 @@ def init_db():
     # import all modules here that might define models so that
     # they will be registered properly on the metadata.  Otherwise
     # you will have to import them first before calling init_db()
-    from models import Department, Employee, Role
-    Base.metadata.drop_all(bind=engine)
-    Base.metadata.create_all(bind=engine)
+    from payment_allocation.models import Customer, Order, Payment, PaymentAllocation
 
-    # Create the fixtures
-    engineering = Department(name='Engineering')
-    db_session.add(engineering)
-    hr = Department(name='Human Resources')
-    db_session.add(hr)
+    load_db = False
+    if load_db:
+        Base.metadata.drop_all(bind=engine)
+        Base.metadata.create_all(bind=engine)
+        # Create the fixtures
+        engineering = Department(name='Engineering')
+        db_session.add(engineering)
+        hr = Department(name='Human Resources')
+        db_session.add(hr)
 
-    manager = Role(name='manager')
-    db_session.add(manager)
-    engineer = Role(name='engineer')
-    db_session.add(engineer)
+        manager = Role(name='manager')
+        db_session.add(manager)
+        engineer = Role(name='engineer')
+        db_session.add(engineer)
 
-    peter = Employee(name='Peter', department=engineering, role=engineer)
-    db_session.add(peter)
-    roy = Employee(name='Roy', department=engineering, role=engineer)
-    db_session.add(roy)
-    tracy = Employee(name='Tracy', department=hr, role=manager)
-    db_session.add(tracy)
-    db_session.commit()
+        peter = Employee(name='Peter', department=engineering, role=engineer)
+        db_session.add(peter)
+        roy = Employee(name='Roy', department=engineering, role=engineer)
+        db_session.add(roy)
+        tracy = Employee(name='Tracy', department=hr, role=manager)
+        db_session.add(tracy)
+        db_session.commit()
